Improve product API stability and align client usage

+++ dient/lv3/src/main/java/com/example/lv3/service/ProductService.java
@@ -6,8 +6,10 @@ import com.example.lv3.dto.response.ProductPageResponse;
 import com.example.lv3.dto.response.ProductResponse;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
+
 public interface ProductService {
-	ProductPageResponse searchProducts(String productName, Long brandId, Long categoryId, Long statusId, Float price, Pageable pageable);
+        ProductPageResponse searchProducts(String productName, Long brandId, Long categoryId, Long statusId, BigDecimal price, Pageable pageable);
 
 	void createProduct(ProductRequest productRequest);
